+++ src/modules/botinfo/commands/index.js
@@ -1,0 +1,3 @@
+export about from './about'
+export help from './help'
+export usage from './usage'