+++ src/modules/index.js
@@ -1,6 +1,15 @@
 import * as osubot from './osubot'
-import * as develop from './develop'
+import * as botinfo from './botinfo'
+import * as utility from './utility'
 
-export const commands = { ...(osubot.commands), ...(develop.commands) }
-export const inits = [ osubot.inits ]
-export const middlewares = [ osubot.middlewares ]
+export const commands = { 
+    ...(osubot.commands),
+    ...(botinfo.commands),
+    ...(utility.commands),
+}
+export const inits = [
+    osubot.inits
+]
+export const middlewares = [
+    osubot.middlewares
+]