+++ src/modules/osubot/canvas/recent.js
@@ -1,9 +1,9 @@
 // Import modules
 import gm from 'gm'
-import fs from 'fs'
-import path from 'path'
+import { copyFileSync, existsSync } from 'fs'
+import { sep } from 'path'
 // Import local files
-import util from './_util'
+import { accuracy, fillNumber, scorify } from './_util'
 import { promisify, promisifyGM, cachepath, assetspath } from './_util'
 import { getAvatar } from './avatar'
 import { res } from '../web'
@@ -27,8 +27,8 @@ export default async function drawRecent(rec, map, stat) {
     const avatarBGDest = `${cachepath}/recentbg/${uid}.jpg`
     const mapFileDest = `${cachepath}/mapfile/${bid}.osu`
     const mods = getMods(rec.enabled_mods)
-    await promisify(fs.copyFile, `${assetspath}/image/userbg/crecent.jpg`, avatarBGDest)
-    if (fs.existsSync(avatarDest) || await getAvatar(uid, avatarDest, avatarLargerDest)) 
+    copyFileSync(`${assetspath}/image/userbg/crecent.jpg`, avatarBGDest)
+    if (existsSync(avatarDest) || await getAvatar(uid, avatarDest, avatarLargerDest)) 
         await promisifyGM(
             gm(avatarBGDest)
             .quality(100)
@@ -36,11 +36,11 @@ export default async function drawRecent(rec, map, stat) {
             .gravity('North')
             .geometry('+0-50')
         )
-    if (!fs.existsSync(bgDest)) {
+    if (!existsSync(bgDest)) {
         try { await res.bgQuery(sid, bgDest) }
-        catch { await promisify(fs.copyFile, `${assetspath}/image/userbg/c${Math.ceil(Math.random() * 5)}.jpg`, bgDest) }
+        catch { copyFileSync(`${assetspath}/image/userbg/c${Math.ceil(Math.random() * 5)}.jpg`, bgDest) }
     }
-    await promisify(fs.copyFile, bgDest, dest)
+    copyFileSync(bgDest, dest)
     await promisifyGM(
         gm(dest)
         .quality(100)
@@ -89,10 +89,10 @@ export default async function drawRecent(rec, map, stat) {
         .fill('#aaa')
         .fontSize(30)
         .drawText(-300, 2, rec.maxcombo + 'x')
-        .drawText(300, 2, util.accuracy(rec) + '%')
+        .drawText(300, 2, accuracy(rec) + '%')
         .fill('#3ad')
         .drawText(-300, 0, rec.maxcombo + 'x')
-        .drawText(300, 0, util.accuracy(rec) + '%')
+        .drawText(300, 0, accuracy(rec) + '%')
         .fontSize(12)
         .fill('#333')
         .drawText(-290, 20, 'max combo')
@@ -108,10 +108,10 @@ export default async function drawRecent(rec, map, stat) {
         .fill('#aaa')
         .drawLine(650, 375, 850, 375)
         .fill('#666')
-        .drawText(-100, 140, util.fillNumber(rec.count300))
-        .drawText(-33, 140, util.fillNumber(rec.count100))
-        .drawText(33, 140, util.fillNumber(rec.count50))
-        .drawText(100, 140, util.fillNumber(rec.countmiss))
+        .drawText(-100, 140, fillNumber(rec.count300))
+        .drawText(-33, 140, fillNumber(rec.count100))
+        .drawText(33, 140, fillNumber(rec.count50))
+        .drawText(100, 140, fillNumber(rec.countmiss))
         .font(`${assetspath}/fonts/Exo2.0-ExtraBold.otf`)
         .fontSize(12)
         .fill('#66a')
@@ -125,11 +125,11 @@ export default async function drawRecent(rec, map, stat) {
         .font(`${assetspath}/fonts/Venera-300.otf`)
         .fontSize(50)
         .fill('#f69')
-        .drawText(0, 0, util.scorify(rec.score))
+        .drawText(0, 0, scorify(rec.score))
         .crop(1000, 500)
     )
     try {
-        if (!fs.existsSync(mapFileDest)) await res.mapFileQuery(bid, mapFileDest)
+        if (!existsSync(mapFileDest)) await res.mapFileQuery(bid, mapFileDest)
         const info = calc(mapFileDest, rec)
         await promisifyGM(
             gm(dest)
@@ -169,5 +169,5 @@ export default async function drawRecent(rec, map, stat) {
             .geometry((padding >= 0 ? '+' : '') + padding + '+170')
         )
     }
-    return 'file://' + process.cwd() + path.sep + dest
+    return 'file://' + process.cwd() + sep + dest
 }