+++ src/modules/osubot/canvas/stat.js
@@ -1,9 +1,9 @@
 // Import modules
 import gm from 'gm'
-import fs from 'fs'
-import path from 'path'
+import { copyFileSync, existsSync } from 'fs'
+import { sep } from 'path'
 // Import local files
-import util from './_util'
+import { scorify, fillNumberReversed, objDiff } from './_util'
 import { promisify, promisifyGM, cachepath, assetspath } from './_util'
 import { getAvatar } from './avatar'
 
@@ -19,7 +19,7 @@ export default async function drawStat(stat, statPrev) {
     const avatarDest = `${cachepath}/avatar/${uid}.jpg`
     const avatarLargerDest = `${cachepath}/avatarl/${uid}.jpg`
     const ranks = ['XH', 'X', 'SH', 'S', 'A']
-    await promisify(fs.copyFile, `${assetspath}/image/userbg/c${Math.ceil(Math.random() * 5)}.jpg`, dest)
+    copyFileSync(`${assetspath}/image/userbg/c${Math.ceil(Math.random() * 5)}.jpg`, dest)
     await promisifyGM(
         gm(dest)
         .quality(100)
@@ -28,7 +28,7 @@ export default async function drawStat(stat, statPrev) {
         .fill('#0005')
         .drawRectangle(30, 20, 400, 390)
     )
-    if (fs.existsSync(avatarDest) || await getAvatar(uid, avatarDest, avatarLargerDest))
+    if (existsSync(avatarDest) || await getAvatar(uid, avatarDest, avatarLargerDest))
         await promisifyGM(
             gm(dest)
             .quality(100)
@@ -47,11 +47,11 @@ export default async function drawStat(stat, statPrev) {
         .fontSize(40)
         .drawText(420, 100, stat.username)
         .fontSize(45)
-        .drawText(30, 700, util.scorify(stat.playcount))
+        .drawText(30, 700, scorify(stat.playcount))
         .drawText(30, 790, stat.accuracy.slice(0, 3 + stat.accuracy.split('.')[0].length) + '%')
-        .drawText(30, 880, util.scorify(stat.ranked_score))
+        .drawText(30, 880, scorify(stat.ranked_score))
         .font(`${assetspath}/fonts/Exo2.0-Bold.otf`)
-        .drawText(30, 610, util.scorify(stat.pp_raw.split('.')[0]) + (stat.pp_raw.split('.')[1] ? ('.' + util.fillNumberReversed(stat.pp_raw.split('.')[1].slice(0, 2), 2)) : '') + 'pp')
+        .drawText(30, 610, scorify(stat.pp_raw.split('.')[0]) + (stat.pp_raw.split('.')[1] ? ('.' + fillNumberReversed(stat.pp_raw.split('.')[1].slice(0, 2), 2)) : '') + 'pp')
         .fontSize(30)
         .drawText(60, 980, stat.count_rank_ssh)
         .drawText(205, 980, stat.count_rank_ss)
@@ -69,13 +69,13 @@ export default async function drawStat(stat, statPrev) {
         .fill('#fff')
         .font(`${assetspath}/fonts/Venera-700.otf`)
         .fontSize(45)
-        .drawText(30, 480, '#' + util.scorify(stat.pp_rank))
+        .drawText(30, 480, '#' + scorify(stat.pp_rank))
         .gravity('NorthEast')
         .fontSize(35)
-        .drawText(70, 480, '#' + util.scorify(stat.pp_country_rank))
+        .drawText(70, 480, '#' + scorify(stat.pp_country_rank))
     )
     if (statPrev) {
-        const diff = util.objDiff(stat, statPrev)
+        const diff = objDiff(stat, statPrev)
         await promisifyGM(
             gm(dest)
             .quality(100)
@@ -111,5 +111,5 @@ export default async function drawStat(stat, statPrev) {
         .composite(`${assetspath}/image/flags/${stat.country}.png`)
         .geometry('+30+455')
     )
-    return 'file://' + process.cwd() + path.sep + dest
+    return 'file://' + process.cwd() + sep + dest
 }