+++ src/modules/utility/commands/index.js
@@ -1,0 +1,2 @@
+export roll from './roll'
+export test from './test'